PiastCode Hackaton
This Hackaton took place in my hometown. We had many issues while coding
this app. This is final form, which in my opinion was the worst app on
whole event. Nevertheless I'm pretty happy with the amount of work I put
to develop this. I'm a humble guy so I would like to thanks my team
which code with me for a 24 hour. Next time will be better.

I want to develop the idea but in new, maybe better, way. Follow my
GitHub to see more in the fututre! :)